refactor: put diary and reading_note types to journal.ts
