+++ src/types/journal.ts
@@ -1,5 +1,15 @@
 import type { Tag } from './tag';
 
+export interface Diary {
+    id: string;
+    text: string | null;
+    date: string;
+    tags: Tag[];
+}
+
+export type DiaryKey = 'Text' | 'Date' | 'TagIds';
+
+
 export interface ReadingNote {
     id: string;
     title: string;
